door push open and close on either direction
