+++ Assets/Scripts/BackDoorOpen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2420a350ce793d48aee0fe19c328fab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
